Add ability to source/import file sans selection
Added the ability to source mel file or import and reload python module.
This sends the current view/file if nothing is selected via the same
command. Added language detection to the command so removed the explicit
language arg. This also allows for a single context menu item which was
added.

Updated readme to match