Update readme and cancel on non-maya language
I'm not sure if it is possible to only show the context menu on MEL and
Python, so it just returns if it's not one of those.

Updating the readme for examples to add to the usersetup.py and indicate
that the current file must be on the user's maya path to have it
actually source/import in maya.